pyproject.toml
[tool.poetry]
name = "cringebot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
"discord.py" = "^1.6.0"
discord-py-slash-command = "^1.0.8"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cringebot = "cringebot:main"
